Split createApp tests
createApp() is responsible for *a lot*, and consequently there are a lot of
tests. Move tests related to registering actions, custom elements, stores and
widgets into separate files, as well as the #realize() implementation.

Share helper functions between the files.
