+++ tests/unit/all.ts
@@ -1,3 +1,8 @@
 import './createApp';
+import './createApp/actions';
+import './createApp/customElements';
+import './createApp/realize';
+import './createApp/stores';
+import './createApp/widgets';
 import './IdentityRegistry';
 import './main';
